feat: reorganise demo code
