+++ src/helpers.tsx
@@ -26,3 +26,7 @@ export function buttonStyle(
       : '$gray5',
   } as const
 }
+
+/** Poor man's tamagui */
+export const $lg = 20
+export const $xl = 30
