--- src/style.ts
@@ -1,3 +1,0 @@
-/** Poor man's tamagui */
-export const $lg = 20
-export const $xl = 30
